requirements.txt
flask
flask-cors
pywebview
requests
pyinstaller

